match: falco

+++ src/app/falco.h
@@ -11,6 +11,6 @@ void falcoDeinit(void);
 void falcoFrameUpdate(void);
 s32 falco_802E51E8(s32, f32, f32, f32);
 s32 falco_802E57C4(void);
-s16 falco_802E5818(void);
+s32 falco_802E5818(void);
 
 #endif // APP_FALCO_H
